# DHC_Backend/DHC_AC_Installer.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DHC AC Installer API",
    "description": "DHC AC Installer 前后端 API 接口规范，与后端 Gin 路由一一对应",
    "version": "2.0.0"
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:19810",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
...
    "/api/installations": {
      "post": {
        "summary": "创建安装任务",
        "description": "根据安装集标识创建异步安装任务，立即返回 installId 供后续轮询",
        "tags": ["Installation"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstallationReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "任务创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateInstallationResp"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResp"
                }
              }
            }
          }
        }
      }
    },
    "/api/installations/{installId}/progress": {
      "get": {
        "summary": "查询安装进度",
        "description": "查询指定安装任务的实时进度，支持按类别过滤",
        "tags": ["Installation"],
        "parameters": [
          {
            "name": "installId",
            "in": "path",
            "description": "安装任务 ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "类别过滤（all 返回全部，或指定 cm/core/weather/map/cars 等）",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回进度快照",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstallationProgressResp"
                }
              }
            }
          },
          "404": {
            "description": "任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResp"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateInstallationReq": {
        "type": "object",
        "properties": {
          "versionId": {
            "type": "string",
            "description": "安装集标识（如 demo-install-v1、demo-resource-verify-v1）"
          },
          "demoSlowProgress": {
            "type": "boolean",
            "description": "是否启用 DEMO 慢速进度模式",
            "default": false
          },
          "demoSlowTotalSeconds": {
            "type": "number",
            "format": "float",
            "description": "慢速模式目标总耗时（秒），范围 [1,300]，默认 20",
            "default": 20
          }
        },
        "required": ["versionId"]
      },
      "CreateInstallationResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "安装任务 ID"
          },
          "versionId": {
            "type": "string",
            "description": "安装集标识"
          },
          "status": {
            "type": "string",
            "enum": ["preparing"],
            "description": "初始状态"
          },
          "startTime": {
            "type": "integer",
            "format": "int64",
            "description": "创建时间（Unix 时间戳，秒）"
          }
        },
        "required": ["id", "versionId", "status", "startTime"]
      },
      "InstallCategoryProgress": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "description": "类别 ID（cm、core、weather、map、cars 等）"
          },
          "categoryName": {
            "type": "string",
            "description": "类别显示名称"
          },
          "status": {
            "type": "string",
            "enum": ["waiting", "active", "completed", "failed"],
            "description": "类别状态"
          },
          "progress": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 100,
            "description": "进度百分比"
          },
          "currentItem": {
            "type": "string",
            "nullable": true,
            "description": "当前处理项名称"
          },
          "totalItems": {
            "type": "integer",
            "description": "总子项数量"
          },
          "completedItems": {
            "type": "integer",
            "description": "已完成子项数量"
          },
          "subProgress": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 100,
            "description": "阶段内子进度（0-100）"
          }
        },
        "required": ["categoryId", "categoryName", "status", "progress"]
      },
      "InstallationProgressResp": {
        "type": "object",
        "properties": {
          "installId": {
            "type": "string",
            "description": "安装任务 ID"
          },
          "status": {
            "type": "string",
            "enum": ["preparing", "installing", "completed", "failed", "interrupted"],
            "description": "总体状态"
          },
          "totalProgress": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 100,
            "description": "总进度百分比"
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InstallCategoryProgress"
            },
            "description": "各类别进度"
          },
          "startTime": {
            "type": "integer",
            "format": "int64",
            "description": "开始时间（Unix 时间戳，秒）"
          },
          "endTime": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "结束时间"
          },
          "error": {
            "type": "string",
            "nullable": true,
            "description": "错误信息"
          }
        },
        "required": ["installId", "status", "totalProgress", "categories", "startTime"]
      },
...
      "ErrorResp": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误消息"
          },
          "detail": {
            "type": "string",
            "nullable": true,
            "description": "详细信息"
          }
        },
        "required": ["error"]
      }
    }
  },
  "security": []
}