# DHC_Frontend/src/api/models/CreateInstallationReq.ts
/* tslint:disable */
/* eslint-disable */
/**
 * DHC AC Installer API
 * DHC AC Installer 前后端 API 接口规范
 *
 * The version of the OpenAPI document: 2.0.0
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CreateInstallationReq {
    versionId: string;
    demoSlowProgress?: boolean;
    demoSlowTotalSeconds?: number;
}

export function CreateInstallationReqFromJSON(json: any): CreateInstallationReq {
    if (json == null) return json;
    return {
        'versionId': json['versionId'],
        'demoSlowProgress': json['demoSlowProgress'],
        'demoSlowTotalSeconds': json['demoSlowTotalSeconds'],
    };
}

export function CreateInstallationReqToJSON(value?: CreateInstallationReq | null): any {
    if (value == null) return value;
    return {
        'versionId': value['versionId'],
        'demoSlowProgress': value['demoSlowProgress'],
        'demoSlowTotalSeconds': value['demoSlowTotalSeconds'],
    };
}